Organize OpenAI unit tests
